# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIGEA REST API",
        "description": "API REST del Sistema Integral de Gestión Educativa Académica (SIGEA). Autenticación mediante tokens Sanctum (header `Authorization: Bearer <token>`).",
        "contact": {
            "name": "Equipo SIGEA",
            "email": "contacto@sigea.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/noticias": {
            "get": {
                "tags": [
                    "Noticias"
                ],
                "summary": "Listar noticias activas",
                "operationId": "acdc7a144931171441c5b1711867cdf7",
                "parameters": [
                    {
                        "name": "desde",
                        "in": "query",
                        "description": "Fecha mínima (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Resultados por página",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Noticias"
                ],
                "summary": "Crear una noticia",
                "operationId": "6edd4c381cc2b94c17b4166d4f8f6688",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "contenido"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "maxLength": 200
                                    },
                                    "contenido": {
                                        "type": "string"
                                    },
                                    "imagen_url": {
                                        "type": "string",
                                        "format": "uri",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Noticia creada"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/noticias/{id}": {
            "get": {
                "tags": [
                    "Noticias"
                ],
                "summary": "Obtener una noticia",
                "operationId": "22f7a75e1b6a99723bcd82807aadcbb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Noticia encontrada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Noticias"
                ],
                "summary": "Actualizar una noticia",
                "operationId": "e9a0ab9b0fb9310dd924cd4a0265fcbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "maxLength": 200
                                    },
                                    "contenido": {
                                        "type": "string"
                                    },
                                    "imagen_url": {
                                        "type": "string",
                                        "format": "uri",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Noticia actualizada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Noticias"
                ],
                "summary": "Eliminar una noticia",
                "operationId": "86e02d017f55d8af043ca15b5252dec3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Noticia eliminada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/notificaciones": {
            "get": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Listar notificaciones recientes del usuario",
                "operationId": "ec57e8c677c5db47f42fbae9c251485f",
                "parameters": [
                    {
                        "name": "no_leidas",
                        "in": "query",
                        "description": "Filtrar solo no leídas",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado con contador de no leídas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/notificaciones/{id}/leida": {
            "post": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Marcar una notificación como leída",
                "operationId": "3df60ee2568a1108a6736895b7b7812c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación actualizada"
                    },
                    "403": {
                        "description": "No pertenece al usuario"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/notificaciones/marcar-todas": {
            "post": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Marcar todas las notificaciones como leídas",
                "operationId": "c1259470ac3aa9a3e4727f8dd515d354",
                "responses": {
                    "200": {
                        "description": "Operación completada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}